more info fixed , custom resource
refactored custom resources to its own model
added custom resource gain
